Move custom templates into the config package
