+++ lib/galaxy/config/template.py
