+++ packages/config/setup.cfg
@@ -33,6 +33,7 @@ include_package_data = True
 install_requires =
     galaxy-util
     boltons
+    jinja2
     pykwalify
     PyYAML
 packages = find:
