+++ packages/util/setup.cfg
@@ -35,7 +35,6 @@ install_requires =
     boltons
     docutils
     importlib_resources
-    jinja2
     MarkupSafe
     packaging
     pycryptodome
