+++ test/unit/app/managers/test_UserManager.py
@@ -217,9 +217,6 @@ class TestUserManager(BaseTestCase):
             == ignore_email_capitalization_user
         )
 
-    def test_register_user_activation(self):
-        pass
-
 
 # =============================================================================
 class TestUserSerializer(BaseTestCase):
